Change type of percentage from int to double Add last block getters in bip47.Wallet

+++ core/src/main/java/org/bitcoinj/wallet/bip47/listeners/BlockchainDownloadProgressTracker.java
@@ -29,5 +29,5 @@ public abstract class BlockchainDownloadProgressTracker extends DownloadProgress
         return isDownloading;
     }
 
-    public abstract int getProgress();
+    public abstract double getProgress();
 }
